upload success

+++ expressApp/apis/listApi.js
@@ -1,3 +1,6 @@
+
+var fs = require('fs');
+
 const listApi = [{
 	url: '/list/getProductList',
 	callback: (req, res)=>{
@@ -15,28 +18,34 @@ const listApi = [{
 
 		let { name } = req.body;
 		let { fileUpload } = req.files;
-
 		console.log(fileUpload);
 
-		var fs = require('fs');
-		var source = fs.createReadStream(fileUpload.path);
+		var { size, type, path } = fileUpload;
 
-		var dest = fs.createWriteStream('D:/upload');
-		source.pipe(dest);
-		source.on('end', function() { 
-			fs.unlinkSync(fileUpload.path);
-		});   //delete
-		source.on('error', function(err) {
-
-		});
+		// var source = fs.createReadStream(fileUpload.path);
+		// var dest = fs.createWriteStream('D:\\0vuejs\\vue-cli\\vue-cli-project\\expressApp\\upload');
+		// source.pipe(dest);
+		// source.on('end', function() { 
+		// 	fs.unlinkSync(fileUpload.path);
+		// });   //delete
+		// source.on('error', function(err) {});
 
 	    setTimeout(function() {
-	    	res.send({
-				success: true,
-				status: 200,
-				msg:'上传成功！'
-			})
-	    }, 2000);
+	    	if(size > 1024 * 1024 * 10) { // 上传的文件大小不能大于10M
+	    		fs.unlinkSync(fileUpload.path);
+	    		res.send({
+					success: false,
+					status: 200,
+					msg:'上传失败！'
+				})
+	    	}else{
+	    		res.send({
+					success: true,
+					status: 200,
+					msg:'上传成功！'
+				})
+	    	}
+	    }, 1000);
 
 	}
 }];
